Fase 3/PREDICTIVA/requirements.txt
# requirements.txt
django
djangorestframework
dj-database-url
python-decouple
psycopg2-binary
matplotlib
tensorflow
scikit-learn
numpy
keras